+Moved slack-message to adapters/slack
