+++ src/adapters/slack/slack-mesage.js
